Restructured to more rational app directory layout.
