+++ imports/api/agent-factory.js
@@ -8,10 +8,10 @@ import humanInterval from 'human-interval';
 
 import { analytics } from "meteor/okgrow:analytics";
 
-import Agent from '/imports/api/agents.js';
-import Transaction from '/imports/models/transaction.js';
+import Agent from '/imports/collections/agents.js';
+import Transaction from '/imports/collections/transactions.js';;
 import Queue from '/imports/api/job-queue.js';
-import Script from '/imports/api/scripts.js';
+import Script from '/imports/collections/scripts.js';
 
 const purgeLimit = humanInterval('1 minute'); // seconds until an agent is considered "dead" and purged
 const agentWatchdogTick = humanInterval('3 seconds'); // check each active agent's status every 2 seconds 
@@ -249,9 +249,12 @@ Queue.define(
     // console.log('Agenda scriptRex transaction:', transaction);
 
     agent.save();
-    job.remove();
+
+
+    // following needs to be moved to an OnSuccess event after Fusion360 finished the job.
+    // FIX!!! This is swallowing jobs!
+    job.remove(); 
     console.log(`## Job ${id} - Job successfully delegated to agent ${agent._id}`);
-    // console.log(agent);
   })
 );
