--- imports/browser/navbar.js
@@ -1,1 +1,0 @@
-// nothing
