--- imports/dashboard.js
@@ -1,5 +1,0 @@
-// export const Dashboard = new Mongo.Collection('dashboard');
-
-// if (!Dashboard.findOne({})) {
-// 	Dashboard.insert('response', {foo: 'bar'});
-// }
