--- imports/demo/index.html
@@ -1,23 +1,0 @@
-<template name="demoLayout">
-  <h1>Fusion Web Demo</h1>
-  <div>
-    <h2>Parameters</h2> {{> parameters_editor }}
-  </div>
-</template>
-
-<template name="parameters_editor">
-  <table>
-    <thead>
-      <th>Name</th>
-      <th>Value</th>
-    </thead>
-    <tbody>
-      {{#each get_params}}
-      <tr>
-        <td> {{ name }} </td>
-        <td>  <input type="text" value={{value}}> </td>
-      </tr>
-      {{/each}}
-    </tbody>
-  </table>
-</template>
