+++ imports/startup/client.js
@@ -1,0 +1,3 @@
+import './both.js';
+
+// import '/imports/api/mixpanel.js'
