+++ imports/startup/server.js
@@ -1,0 +1,3 @@
+import './both.js';
+
+import '/imports/publications/publications.js'
